+++ Braintree/src/main/java/com/braintreepayments/api/models/CardBuilder.java
@@ -134,6 +134,7 @@ public class CardBuilder extends BaseCardBuilder<CardBuilder> implements Parcela
                 "      brand" +
                 "      expirationMonth" +
                 "      expirationYear" +
+                "      cardholderName" +
                 "      last4" +
                 "      binData {" +
                 "        prepaid" +